added camera class and updated config
The camera class has the ability to calibrate the camera. The config serves as a default value. Fine tune it in the config. only the quitting of the calibration does not seem to work right at the moment.

The camera gets a config class with the values for camera id, resolution, hue, saturation and value.