Using object detection
Update to program: I have trained my own data set on yolov4tiny using my own photos and some images i found online. I have now implemented this into the code. It detects the plates from the image. However it does not draw the bounding box around it correctly but i'll fix that.

There are a lot on clean ups to do. Like making sure the path to the darknet files is not hard coded anymore (oops... :grinning: )
it might also be time to run cppcheck to check for problems.

A small summary of what has happend:

- Removed Calibration (Not needed for object detection, could have know... -_- )
- Trained my own dataset
- Implemented darknet in my project
- Expanded the config file
- Fixed bug regarding the creation of the camera object.