Added OCR and better detection
The OCR is not perfect and im still finetuning it. It does not always the correct characters. I also need to do better filtering (using a regex maybe?)
The filter function is not complete yet.

I also added a Non-Maximum Suppression (NMS) Threshold to help with the multiple detection's of plates and the drawing of multiple boxes.

There also is a lot of optimization to do but that is for a later date. I first want everything to work so i know where i can optimize.