Added better text filtering
Using regex's to check for licence plates in the string.
This means that if a character is falsely detected, for example:
"BXX-XX-99" it returns the string as "XX-XX-99"

I also renamed the filter function to remove_dash. The OCR class already sets the chars we need. The remove_dash function is purely for the database. (It has the plates stored as "XXXX99"

I also might move these functions somewhere else... this would be for the cleanup and optimization part.