Warped ROI
Still needs to be leveled.

Also made the change to exchange the letter 'o' to a zero since it confused the OCR model.

This means trailers are not detectable
(Maybe it is possible to exchange the O based on other parameters of the string? )

There also seems to be a bug with the plates getting detected twice?